+++ src/Declaration/class.DeclarationsCtrl.php
@@ -7,6 +7,7 @@ use ilObject;
 use ilSrSelfDeclarationPlugin;
 use ilUIPluginRouterGUI;
 use srag\DIC\SrSelfDeclaration\DICTrait;
+use srag\Plugins\SrSelfDeclaration\Config\Config;
 use srag\Plugins\SrSelfDeclaration\Utils\SrSelfDeclarationTrait;
 
 /**
@@ -30,6 +31,10 @@ class DeclarationsCtrl
     const LANG_MODULE = "declarations";
     const PLUGIN_CLASS_NAME = ilSrSelfDeclarationPlugin::class;
     const TAB_LIST_DECLARATIONS = "list_declarations";
+    /**
+     * @var Config
+     */
+    protected $config;
     /**
      * @var ilObject
      */
@@ -80,6 +85,8 @@ class DeclarationsCtrl
 
         self::dic()->ctrl()->saveParameter($this, self::GET_PARAM_REF_ID);
 
+        $this->config = self::srSelfDeclaration()->configs()->getConfig($this->obj->getId());
+
         $this->setTabs();
 
         $next_class = self::dic()->ctrl()->getNextClass($this);
@@ -118,7 +125,7 @@ class DeclarationsCtrl
     {
         self::dic()->tabs()->activateTab(self::TAB_LIST_DECLARATIONS);
 
-        $table = self::srSelfDeclaration()->declarations()->factory()->newTableBuilderInstance($this, $this->obj);
+        $table = self::srSelfDeclaration()->declarations()->factory()->newTableBuilderInstance($this, $this->obj, $this->config);
 
         self::output()->output($table, true);
     }