+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
 
+## [1.1.0]
+- Declaration effort
+- Fix table pagination
+
 ## [1.0.0]
 - First version